added view controller extensions
