+++ free-tunes/Controllers/LoginViewController.swift
@@ -78,16 +78,4 @@ extension LoginViewController: UITextFieldDelegate {
     }
 }
 
-extension UIViewController {
 
-    func displayErrorAlert(alertTitle: String,
-                           alertMessage: String,
-                           alertActionTitle: String,
-                           alertDelegate: UIViewController) {
-        let alert = UIAlertController(title: alertTitle,
-                                      message: alertMessage,
-                                      preferredStyle: UIAlertController.Style.alert)
-        alert.addAction(UIAlertAction(title: alertActionTitle, style: UIAlertAction.Style.default, handler: nil))
-        alertDelegate.present(alert, animated: true, completion: nil)
-    }
-}
