+++ free-tunes/Utilities/CustomViewController.swift
@@ -1,0 +1,22 @@
+//
+//  CustomViewController.swift
+//  free-tunes
+//
+//  Created by Tshwarelo Mafaralala on 2022/03/07.
+//
+
+import UIKit
+
+extension UIViewController {
+
+    func displayErrorAlert(alertTitle: String,
+                           alertMessage: String,
+                           alertActionTitle: String,
+                           alertDelegate: UIViewController) {
+        let alert = UIAlertController(title: alertTitle,
+                                      message: alertMessage,
+                                      preferredStyle: UIAlertController.Style.alert)
+        alert.addAction(UIAlertAction(title: alertActionTitle, style: UIAlertAction.Style.default, handler: nil))
+        alertDelegate.present(alert, animated: true, completion: nil)
+    }
+}
